Deployed to main net
Set a total token supply of 300k. I plan to give 1k to each person on
the cohort and then there will be around 200k for miscellaneous
projects. I am happy to decide on how these tokens are used in a
democratic process.

Set the owner as my main Ethereum account. For now I will be the sole
controller of the contract.